// src/lib/types/driver.ts
export interface Driver {  
  firstName: string;  
  lastName: string;  
  curp: string;  
  street: string;  
  city: string;  
  state: string;  
  monthlySalary: number;  
  licenseNumber: string;  
  registrationDate: string;
}

export interface DriverPublic {  
  firstName: string;  
  lastName: string;  
  curp: string;  
  street: string;  
  city: string;  
  state: string;  
  monthlySalary: number;  
  licenseNumber: string;  
  registrationDate: string;  
} 